Projection

Projection

Many NEP-solvers are based on a computation of a projected problem, i.e., if $V,W\in\mathbb{R}^{n\times p}$ we need to solve the (smaller) NEP

\[W^HM(λ)Vz=0\]

This is sometimes called a nonlinear Rayleigh-Ritz procedure, or a direct projection. These are inner solvers for many problems.

NEP-PACK provides a framework to handle projected problems and inner solves. You can in principle use any of the NEP-solvers to solve a projected problem. As a user, this is specified in the inner_solver_method keyword argument in those NEP-solvers that use a direct projection. By default DefaultInnerSolver is used.

If you wish to use the infinite Arnoldi method to handle the project solves in the nonlinear Arnoldi method, you can do the following:

julia> nep=nep_gallery("dep0_tridiag");
julia> λ,v=nlar(nep,neigs=1,inner_solver_method=IARInnerSolver(),logger=1);
Using inner solver IARInnerSolver(1.0e-13, 80, :ones, false, NonlinearEigenproblems.NEPSolver.iar)
iter 1 err:0.05095382004494062 λ=0.7579134426195271 - 0.03707164055891316im
iter 2 err:0.00031997693290503965 λ=-0.00010049358638757657 + 0.0001763732030940319im
iter 3 err:6.563177508431498e-6 λ=-0.0005335154073888051 - 4.498082881742902e-6im
iter 4 err:8.037612383023366e-9 λ=-0.0005259618179586685 + 2.806438064753968e-9im
iter 5 err:3.386041718599221e-11 λ=-0.0005259683064505526 + 1.3119844096548209e-11im
iter 6 err:3.4499779767886924e-13 λ=-0.0005259682927702825 - 1.0404412824030558e-13im
iter 7 err:5.365662696809372e-15 λ=-0.0005259682929434785 - 2.0247938561528697e-17im
****** 1 converged to eigenvalue: -0.0005259682929434785 - 2.0247938561528697e-17im errmeasure:5.365662696809372e-15

The logging of the inner solver is controlled by the kwarg inner_logger. This produces very verbose output illustrating also the convergence of the inner solve:

julia> λ,v=nlar(nep,neigs=1,inner_solver_method=IARInnerSolver(),logger=1,inner_logger=1);
Using inner solver IARInnerSolver(1.0e-13, 80, :ones, false, NonlinearEigenproblems.NEPSolver.iar)
-
--
---
----
-----
------
-------
--------
---------
----------
-----------
------------
=------------
+-------------
iter 1 err:0.04860520921206162 λ=0.8437634284420165 + 0.005742468974957178im
-
--
---
+---
iter 2 err:0.0008771218464072076 λ=-0.00065275394814732 - 0.0008601482370586537im
-
--
---
...

Inner solvers

The inner solvers inherit from InnerSolver. The following inner solvers are available by default.

NonlinearEigenproblems.NEPSolver.NewtonInnerSolver — Type.

struct NewtonInnerSolver
function NewtonInnerSolver(;tol=1e-13,maxit=80,starting_vector=:Vk,
                           newton_function=augnewton)

Uses a Newton-like method to solve the inner problem, with tolerance, and maximum number of iterations given by tol and maxit. The starting vector can be :ones, :randn, or :Vk. The value :Vk specifies the use of the outer NEP-solver keyword argument (Vk). This is typically the previous iterate in the outer method.

The kwarg newton_function, specifies a Function which is called. The type supports augnewton, newton, resinv quasinewton, newtonqr. In principle it can be any method which takes the same keyword arguments as these methods.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.IARInnerSolver — Type.

struct IARInnerSolver
function IARInnerSolver(;tol=1e-13,maxit=80,
           starting_vector=:ones,normalize_DEPs=:auto,
           iar_function=iar)

Uses iar, tiar or iar_chebyshev to solve the inner problem, with tolerance, and maximum number of iterations given by tol and maxit. The starting vector can be :ones or :randn. The iar_function can be iar, tiar or iar_chebyshev (or any function taking the same parameters as input). normalize_DEPs determines if the we should carry out precomputation and make sure the projection of a DEP is again a DEP (which can speed up performance). It can take the value true, false or :auto. :auto sets it to true if we use the iar_chebyshev solver.

The kwarg iar_function, specifies a Function which is called. Examples of functions are iar and iar_chebyshev. It can be any NEP-solver which takes the same keyword arguments as these methods.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.IARChebInnerSolver — Function.

function IARChebInnerSolver(;tol=1e-13,maxit=80,starting_vector=:ones,
                            normalize_DEPs=true)

Uses iar_chebyshev to solve the inner problem. See IARInnerSolver for keyword argument documentation.

See also: IARInnerSolver, InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.ContourBeynInnerSolver — Type.

struct ContourBeynInnerSolver <: InnerSolver
function ContourBeynInnerSolver(;tol=sqrt(eps(real(Float64))),
                                radius=:auto,N=1000)

Uses contour_beyn to solve the inner problem, with radius and number of quadrature nodes, given by radius and n. If the variable radius is set to :auto, the integration radius will be automatically by using the eigenvalue approximations specified by the outer solver.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.PolyeigInnerSolver — Type.

struct PolyeigInnerSolver <: InnerSolver
function PolyeigInnerSolver()

Specifies the use of polyeig to solve the inner problem. This is intended for NEPs of the type PEP.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.SGIterInnerSolver — Type.

struct SGIterInnerSolver <: InnerSolver

Uses sgiter to solve the inner problem.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.NleigsInnerSolver — Type.

struct NleigsInnerSolver <: InnerSolver
function NleigsInnerSolver(;Σ= :auto,nodes =:auto, tol=1e-6 )

Uses nleigs to solve the inner problem, in the region Σ with shifts nodes and with tolerance tol. If the variable Σ is set to :auto, the region Σ will be set by using the eigenvalues approximations. See nleigs for description of parameters.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.DefaultInnerSolver — Type.

struct DefaultInnerSolver <: InnerSolver

Dispatches a version of inner_solve based on the type of the NEP provided.

See also: InnerSolver, inner_solve

Advanced usage

You can define your own inner solver by inheriting from InnerSolver and implementing the function inner_solve. Since the inner_solve obtains information from the solver via keyword arguments, you need to end your method signature with kwargs...).

NonlinearEigenproblems.NEPSolver.InnerSolver — Type.

abstract type InnerSolver

Structs inheriting from this type are used to solve inner problems in an inner-outer iteration.

The InnerSolver objects are passed to the NEP-algorithms, which uses it to dispatch the correct version of the function inner_solve. Utilizes existing implementations of NEP-solvers and inner_solve acts as a wrapper to these.

Example

There is a DefaultInnerSolver that dispatches an inner solver based on the provided NEP. However, this example shows how you can force nlar to use the IARInnerSolver for a PEP.

julia> nep=nep_gallery("pep0", 100);
julia> λ,v = nlar(nep, inner_solver_method=NEPSolver.IARInnerSolver(), neigs=1, num_restart_ritz_vecs=1, maxit=70, tol=1e-8);
julia> norm(compute_Mlincomb(nep,λ[1],vec(v)))
8.68118417430353e-9

See also: inner_solve, DefaultInnerSolver, NewtonInnerSolver, PolyeigInnerSolver, IARInnerSolver, IARChebInnerSolver, SGIterInnerSolver, ContourBeynInnerSolver

NonlinearEigenproblems.NEPSolver.inner_solve — Function.

inner_solve(is::InnerSolver,T_arit,nep;kwargs...)

Solves the projected linear problem with solver specied with is. This is to be used as an inner solver in an inner-outer iteration. T specifies which method to use. The most common choice is DefaultInnersolver. The function returns (λv,V) where λv is an array of eigenvalues and V a matrix with corresponding vectors. The struct T_arit defines the arithmetics used in the outer iteration and should prefereably also be used in the inner iteration.

Different inner_solve methods take different kwargs. These are standardized kwargs:

neigs: Number of wanted eigenvalues (but less or more may be returned)

σ: target specifying where eigenvalues

λv, V: Vector/matrix of guesses to be used as starting values

j: the jth eigenvalue in a min-max characterization

tol: Termination tolarance for inner solver

inner_logger: Determines how the inner solves are logged. See Logger for further references

Rayleigh functional computation

NonlinearEigenproblems.NEPSolver.compute_rf — Function.

compute_rf(eltype::Type,nep::NEP,x,inner_solver::InnerSolver;
           y=x, target=zero(T), λ0=target,TOL=eps(real(T))*1e3,max_iter=10)

Computes the Rayleigh functional of the nep, i.e., computes a vector $Λ$ of values $λ$ such that $y^TM(λ)x=0$, using the procedure specified in inner_solver. The default behaviour consists of a scalar valued Newton-iteration, and the returned vector has only one element.

The given eltype<:Number is the type of the returned vector.

Example

This uses iar to solve the (scalar) nonlinear problem.

julia> nep=nep_gallery("dep0");
julia> x=ones(size(nep,1));
julia> s=compute_rf(ComplexF64,nep,x,IARInnerSolver())[1]; # Take just first element
0.6812131933795565 + 0.0im
julia> x'*compute_Mlincomb(nep,s,x)
1.7763568394002505e-15 + 0.0im

Projection types

Several methods for NEPs are based on forming a smaller NEP, which we will refer to as a projection:

\[N(λ)=W^HM(λ)V,\]

where $V,W\in\mathbb{C}^{n\times p}$ and the corresponding projected problem

\[N(λ)u=0.\]

Types

NEPs for which this projection can be computed inherit from ProjectableNEP.

NonlinearEigenproblems.NEPTypes.ProjectableNEP — Type.

abstract ProjectableNEP <: NEP

A ProjectableNEP is a NEP which can be projected, i.e., one can construct the problem $W'*M(λ)Vw=0$ with the Proj_NEP. A NEP which is of this must have the function create_proj_NEP(orgnep::ProjectableNEP) implemented. This function must return a Proj_NEP.

See also: set_projectmatrices!.

Example:

julia> nep=nep_gallery("dep0");
julia> typeof(nep)
DEP{Float64,Array{Float64,2}}
julia> isa(nep,ProjectableNEP)
true
julia> projnep=create_proj_NEP(nep);
julia> e1 = Matrix(1.0*I,size(nep,1),1);
julia> set_projectmatrices!(projnep,e1,e1);
julia> compute_Mder(nep,3.0)[1,1]
-2.315345215259418
julia> compute_Mder(projnep,3.0)
1×1 Array{Float64,2}:
 -2.315345215259418

The result of the projection is represented in a Proj_NEP.

NonlinearEigenproblems.NEPTypes.Proj_NEP — Type.

abstract type Proj_NEP <: NEP

Proj_NEP represents a projected NEP. The projection is defined as the NEP

\[N(λ)=W^HM(λ)V\]

where $M(λ)$ is a base NEP and W and V rectangular matrices representing a basis of the projection spaces. Instances are created with create_proj_NEP. See create_proj_NEP for examples.

Any Proj_NEP needs to implement two functions to manipulate the projection:

set_projectmatrices!: Set matrices W and V

expand_projectmatrices!: Effectively expand the matrices W and V with one column.

One explicit instance is the Proj_SPMF_NEP.

NonlinearEigenproblems.NEPTypes.Proj_SPMF_NEP — Type.

struct Proj_SPMF_NEP <: Proj_NEP

This type represents the (generic) way to project NEPs which are AbstractSPMF. See examples in create_proj_NEP.

Associated functions

You can create a projected NEP with create_proj_NEP:

NonlinearEigenproblems.NEPTypes.create_proj_NEP — Function.

pnep=create_proj_NEP(orgnep::ProjectableNEP[,maxsize [,T]])

Create a NEP representing a projected problem $N(λ)=W^HM(λ)V$, where the base NEP is represented by orgnep. The optional parameter maxsize::Int determines how large the projected problem can be and T is the Number type used for the projection matrices (default ComplexF64). These are needed for memory preallocation reasons. Use set_projectmatrices! and expand_projectmatrices! to specify projection matrices $V$ and $W$.

Example:

The following example illustrates that a projection of a NEP is also a NEP and we can for instance call compute_Mder on it:

julia> nep=nep_gallery("pep0")
julia> V=Matrix(1.0*I,size(nep,1),2);
julia> W=Matrix(1.0*I,size(nep,1),2);
julia> pnep=create_proj_NEP(nep);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,3.0)
2×2 Array{Complex{Float64},2}:
 -2.03662+0.0im   13.9777+0.0im
 -1.35069+0.0im  -13.0975+0.0im
julia> W'*compute_Mder(nep,3.0)*V  # Gives the same result
2×2 Array{Float64,2}:
 -2.03662   13.9777
 -1.35069  -13.0975

If you know that you will only use real projection matrices, you can specify this in at the creation:

julia> pnep=create_proj_NEP(nep,2,Float64);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,3.0)
2×2 Array{Float64,2}:
 -2.03662   13.9777
 -1.35069  -13.0975

NonlinearEigenproblems.NEPTypes.set_projectmatrices! — Function.

set_projectmatrices!(pnep::Proj_NEP,W,V)

Set the projection matrices for the NEP to W and V, i.e., corresponding the NEP: $N(λ)=W^HM(λ)V$. See also create_proj_NEP.

Example

This illustrates if W and V are vectors of ones, the projected problem becomes the sum of the rows and columns of the original NEP.

julia> nep=nep_gallery("pep0")
julia> pnep=create_proj_NEP(nep);
julia> V=ones(200,1);  W=ones(200,1);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,0)
1×1 Array{Complex{Float64},2}:
 48.948104019482756 + 0.0im
julia> sum(compute_Mder(nep,0),dims=[1,2])
1×1 Array{Float64,2}:
 48.948104019482955

NonlinearEigenproblems.NEPTypes.expand_projectmatrices! — Function.

expand_projectmatrices!(nep::Proj_SPMF_NEP, Wnew, Vnew)

The projected NEP is updated by adding the last column of Wnew and Vnew to the basis. Note that Wnew and Vnew contain also the "old" basis vectors. See also create_proj_NEP

Example:

In the following example you see that the expanded projected problem has one row and column more, and the leading subblock is the same as the smaller projected NEP.

julia> nep=nep_gallery("pep0"); n=size(nep,1);
julia> V=Matrix(1.0*I,n,2); W=Matrix(1.0*I,n,2);
julia> pnep=create_proj_NEP(nep);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,0)
2×2 Array{Complex{Float64},2}:
 0.679107+0.0im   -0.50376+0.0im
 0.828413+0.0im  0.0646768+0.0im
julia> Vnew=[V ones(n)]
julia> Wnew=[W ones(n)]
julia> expand_projectmatrices!(pnep,Wnew,Vnew);
julia> compute_Mder(pnep,0)
3×3 Array{Complex{Float64},2}:
 0.679107+0.0im   -0.50376+0.0im  -12.1418+0.0im
 0.828413+0.0im  0.0646768+0.0im   16.3126+0.0im
 -17.1619+0.0im   -10.1628+0.0im   48.9481+0.0im