LinSolver

LinSolvers

Most NEP-algorithms need to solve the linear system associated with M(λ). We provide an interface to specify which solver to use or define your own solver.

NonlinearEigenproblems.LinSolvers.LinSolver — Type.

abstract type LinSolver

Structs inheriting from this type are able to solve linear systems associated with a NEP, for a specific λ-value. The most common are direct solvers such as DefaultLinSolver, BackslashLinSolver and iterative solvers such as GMRESLinSolver.

The LinSolver objects are usually created by the NEP-algorithms through creator functions, which are passed as parameters.

Example

The most common usecase is that you want to pass a linsolvercreator-function as parameter to the NEP-algorithm. This example shows how you can solvers based on backslash or factorize(). In the example, BackslashLinSolver does not exploit that the system matrix remains the same throughout the algorithm and is therefore slower.

julia> nep=nep_gallery("qdep0");
julia> using BenchmarkTools
julia> v0=ones(size(nep,1));
julia> @btime λ,v=quasinewton(nep,λ=-1,v=v0, linsolvercreator=default_linsolvercreator);
  199.540 ms (4929 allocations: 59.83 MiB)
julia> @btime λ,v=quasinewton(nep,λ=-1,v=v0, linsolvercreator=backslash_linsolvercreator);
  1.632 s (6137 allocations: 702.85 MiB)

Example

The LinSolvers are constructed for extendability. This example creates our own LinSolver which uses an explicit formula for the inverse if the NEP has dimension 2x2.

Create the types and a creator.

julia> using LinearAlgebra
julia> struct MyLinSolver <: LinSolver
   M::Matrix{ComplexF64}
end
julia> function my_linsolvercreator(nep,λ)
   M=compute_Mder(nep,λ);
   return MyLinSolver(M);
end

Explicit import lin_solve to show how to solve a linear system.

julia> import NonlinearEigenproblems.LinSolvers.lin_solve;
julia> function lin_solve(solver::MyLinSolver,b::AbstractVecOrMat;tol=0)
   M=solver.M;
   invM=(1/(det(M)))*[M[2,2] -M[1,2];-M[2,1] M[1,1]]
   return invM*b
end
julia> nep=SPMF_NEP([[1.0 3.0; 4.0 5.0], [2.0 1.0; -1 2.0]], [S->S^2,S->exp(S)])
julia> λ,v=quasinewton(nep,λ=-1,v=[1;1],linsolvercreator=my_linsolvercreator);

See also: lin_solve, DefaultLinSolver, default_linsolvercreator, BackslashLinSolver, backslash_linsolvercreator, GMRESLinSolver, gmres_linsolvercreator

source

NonlinearEigenproblems.LinSolvers.lin_solve — Function.

lin_solve(solver::LinSolver, b::AbstractVecOrMat; tol=0)

This function solves the linear system represented in solver::LinSolver with a right-hand side b. The tol kwarg is controlling how accurate the linear system needs to be solved. A NEP-algorithm will call this solver every time a linear system associated with M(λ) needs to be solved.

This function must be overloaded if a user wants to define their own way of solving linear systems. See LinSolver for examples.

source

NonlinearEigenproblems.LinSolvers.DefaultLinSolver — Type.

struct DefaultLinSolver <: LinSolver

This represents the linear solver associated with julia factorize(). See LinSolver and default_linsolvercreator for examples.

source

NonlinearEigenproblems.LinSolvers.default_linsolvercreator — Function.

default_linsolvercreator(nep::NEP, λ; umfpack_refinements = 2)

Create a linear solver of type DefaultLinSolver for the NEP evaluated in point λ. For sparse matrices (the underlying solver is usually UMFPACK) the maximum number of iterative refinements can be changed to trade accuracy for performance with the parameter umfpack_refinements. UMFPACK defaults to a maximum of 2 iterative refinements.

For examples see LinSolver.

See also: DefaultLinSolver.

source

NonlinearEigenproblems.LinSolvers.BackslashLinSolver — Type.

struct BackslashLinSolver <: LinSolver

This represents a linear solver corresponding to the backslash operator (no pre-factorization).

See also: LinSolver and backslash_linsolvercreator

source

NonlinearEigenproblems.LinSolvers.backslash_linsolvercreator — Function.

backslash_linsolvercreator(nep::NEP, λ)

Create a linear solver of type BackslashLinSolver evaluated in λ.

See also: LinSolver, BackslashLinSolver

source

NonlinearEigenproblems.LinSolvers.GMRESLinSolver — Type.

struct GMRESLinSolver <: LinSolver

This represents a solver done with the julia GMRES implementation.

See also: LinSolver, gmres_linsolvercreator

source

NonlinearEigenproblems.LinSolvers.gmres_linsolvercreator — Function.

gmres_linsolvercreator(nep::NEP, λ, kwargs=())

Create a linear solver of type GMRESLinSolver. The kwargs are passed as parameter to Julia-built-in-GMRES.

See also: LinSolver, GMRESLinSolver

source

EigSolvers

Some NEP-algorithms need to solve an associated linear eigenvalue problem. associated with M(λ). You will likely only need the native eigensolvers in Julia. Nevertheless, we provide an interface to specify which solver to use or define your own solver.

NonlinearEigenproblems.LinSolvers.EigSolver — Type.

abstract type EigSolver

Structs inheriting from this type are able to solve linear eigenvalue problems arising in certain methods, such as, e.g., mslp, sgiter, and polyeig.

The EigSolver objects are passed as types to the NEP-algorithms, which uses it to dispatch the correct version of the function eig_solve.

Example

The most common usecase is that you do not want to specify anything in particular, since the DefaultEigSolver will use a dense or a sparse method depending on you problem. However, this example shows how you can force mslp to use the sparse solver.

julia> nep=nep_gallery("qdep0");
julia> λ,v = mslp(nep, eigsolvertype=NativeEigSSolver);
julia> norm(compute_Mlincomb(nep,λ,v))
1.0324139764567768e-15

Example

The EigSolvers are constructed for extendability. As an illustartion this example creates a naive EigSolver which casts the problem to a standard linear eigenproblem and calls the built-in function to solve it.

Create the types and a creator.

julia> struct MyEigSolver <: EigSolver
   A
   E
   function MyEigSolver(A,E)
      return new(A,E)
   end
end

julia> import NonlinearEigenproblems.LinSolvers.eig_solve;
julia> function eig_solve(solver::MyEigSolver;nev = 1, target = 0)
   M = solver.E \ solver.A
   eig = eigen(M)
   i = argmin(abs.(eig.values))
   return eig.values[i], eig.vectors[:,i]
end
julia> nep=nep_gallery("dep0", 50);
julia> λ,v = mslp(nep, eigsolvertype=MyEigSolver, tol=1e-5);
julia> norm(compute_Mlincomb(nep,λ,v))
3.0777795031319117e-10

See also: eig_solve, DefaultEigSolver, NativeEigSolver, NativeEigSSolver, eig_solve

source

NonlinearEigenproblems.LinSolvers.eig_solve — Function.

eig_solve(solver::EigSolver; [nev,] [target,])

This function solves the linear eigenvalue problem represented in solver::EigSolver. The nev kwarg is controlling the number of eigenvalues aimed for, and target specifies around which point the eigenvalues are computed. The former has a defalut value equalt to the seize of the problem, and the latter has a defalut value 0.

Return values are of the form (Vector, Matrix) where the former contains the eigenvalues and the latter the eigenvectors.

This function must be overloaded if a user wants to define their own way of solving linear eigenvalue problems. See EigSolver for examples.

source

NonlinearEigenproblems.LinSolvers.DefaultEigSolver — Type.

struct DefaultEigSolver <: EigSolver

A linear eigenvalueproblem solver that calls checks for sparsity and accordingly assigns an appropriate solver.

See also: EigSolver, eig_solve, NativeEigSolver, NativeEigSSolver

source

NonlinearEigenproblems.LinSolvers.NativeEigSolver — Type.

struct NativeEigSolver <: EigSolver

A linear eigenvalueproblem solver that calls Julia's in-built eigen()

Constructed as NativeEigSolver(A, [B,]), and solves the problem

\[Ax = λBx\]

The paramter B is optional an default is indentity, for which a standard linear eigenproblem is solved.

See also: EigSolver and eig_solve

source

NonlinearEigenproblems.LinSolvers.NativeEigSSolver — Type.

struct NativeEigSSolver <: EigSolver

A linear eigenvalueproblem solver for large and sparse problems that calls Julia's in-built eigs()

Constructed as NativeEigSSolver(A, [B,]), and solves the problem

\[Ax = λBx\]

The paramter B is optional an default is indentity, for which a standard linear eigenproblem is solved.

See also: EigSolver and eig_solve

source