Logger

Logger

This is description of the logger functionality in NEP-PACK.

** TODO **

As a user

NonlinearEigenproblems.NEPSolver.Logger — Type.

abstract type Logger ; end

The type represents a way to log information throughout the algorithms, as well as a storage. The most common Loggers are PrintLogger and ErrorLogger.

As a method developer you want to use push_info! and push_iteration_info!.

See also: PrintLogger and ErrorLogger.

source

NonlinearEigenproblems.NEPSolver.PrintLogger — Type.

struct PrintLogger <: Logger ;

When you use this logger, you will obtain printouts in stdout, no other logging. The displaylevel parameter specified, how much should be printed.

source

NonlinearEigenproblems.NEPSolver.ErrorLogger — Type.

struct ErrorLogger <: Logger

When you use this logger, the error of push_iteration_info!-calls will be stored in logger.errs. It can also print to stdout, if displaylevel is set to a value greater than zero.

source

As a solver developer

NonlinearEigenproblems.NEPSolver.push_info! — Function.

function push_info!(logger, [level,] v; continues::Bool=false)

Pushes a string v to the logger. If continues=true, the next push_info! (or push_iteration_info!) is connected to this, e.g. line-feed will be omitted.

source

NonlinearEigenproblems.NEPSolver.push_iteration_info! — Function.

function push_iteration_info!(logger, [level,] iter; kwargs)

Pushes information about a specific iteration iter. The supported keyword arguments include λ, err and v.

source