Measuring the error

Measuring the error

All iterative algorithms need some form of termination criteria. In NEP-PACK, all NEP-solvers provide the possibility to specify the desired tolerance, as well as how the error is measured or estimated. The tolerance is specified in the kwarg tol (which is a real number) and the way to measure the error is given in errmeasure.

Standard usage

NEP-PACK comes with several ways to measure errors for many NEP-types.

errmeasure=ResidualErrmeasure: The error is estimated by the use of the residual norm:

\[\mathrm{err}=\frac{\|M(λ)v\|}{\|v\|}.\]

errmeasure=BackwardErrmeasure: The error is estimated by using the backward error bounds. This error measure will not work for all NEPs. An implementation is provided for any AbstractSPMF. If your NEP is an AbstractSPMF with terms:
\[M(λ)=A_1f_1(λ)+\cdots+A_mf_m(λ)\]
the error will be estimated by
\[\mathrm{err}=\frac{\|M(λ)v\|}{\|v\|}\frac{1}{\|A_1\|_F|f_1(λ)|+\cdots+\|A_m\|_F|f_m(λ)|}.\]
In other words, the BackwardErrmeasure is a weighting of the ResidualErrmeasure.

errmeasure=DefaultErrmeasure: When this errmeasure is specified, NEP-PACK tries to determine a error measure for you. In general, BackwardErrmeasure will be preferred if possible. This behavior may change in future versions of NEP-PACK.

Example: Most NEP-solvers take the errmeasure as an kwarg.

julia> nep=nep_gallery("qdep0");
julia> # Solve the problem to residual norm 1e-8
julia> (λ,v)=mslp(nep,errmeasure=ResidualErrmeasure,tol=1e-8)
julia> norm(compute_Mlincomb(nep,λ,v))/norm(v) # It's smaller than tol?
3.503700819937386e-9
julia> nep isa AbstractSPMF # Is it an AbstractSPMF so we can use BackwardErrmeasure?
true
julia> (λ,v)=mslp(nep,errmeasure=BackwardErrmeasure,tol=1e-10)
julia> factor=abs(fv[1](λ))*norm(Av[1])+
     abs(fv[2](λ))*norm(Av[2])+abs(fv[3](λ))*norm(Av[3]);
julia> norm(compute_Mlincomb(nep,λ,v))/(norm(v)*factor)
1.659169482386331e-11

User defined error measure

There are two ways that a user can specify how to measure the error.

User defined error 1: Function handle

The user can provide a function handle which is called to obtain the error. The errmeasure can be a function, which takes two parameters as input (λ,v) and returns the error (or estimate thereof).

The most common situation is that you want to report the error (as a function of iteration) with a reference solutions. If we want to get a very accurate approximation of the true error, we can run the algorithm twice, and the second time we run the algorithm we use the result of the first run as a reference.

julia> nep=nep_gallery("qdep0");
julia> v0=ones(size(nep,1));
julia> (λref,v)=resinv(nep,v=v0,λ=230^2+1im,logger=1);
julia> myerrmeasure = (λ,v) -> abs(λ-λref)/abs(λ);
julia> (λ,v)=resinv(nep,v=v0,λ=250^2+1im,logger=1,tol=1e-10,errmeasure=myerrmeasure);
Iteration:  1 errmeasure:1.274091618457501296e-01
Iteration:  2 errmeasure:9.535794095609478882e-01
...
Iteration: 49 errmeasure:1.269396691930517923e-10
Iteration: 50 errmeasure:7.608430406801784718e-11

User defined error 2: A user defined type

Due to the multiple dispatch and handling of types in Julia, code may run faster if one uses types instead of function handles. It is possible to do the same simulation as above with a user defined type.

You first need to define a new type

julia> struct RefErrmeasure <: Errmeasure; nep::NEP; end

The error measure should then provided in the function estimate_error:

julia> v0=ones(size(nep,1));
julia> (λref,v)=resinv(nep,v=v0,λ=230^2+1im,logger=1);
julia> function NonlinearEigenproblems.estimate_error(e::RefErrmeasure,λ,v)
         return abs(λ-λref)/abs(λ);
       end
julia> (λ,v)=resinv(nep,v=v0,λ=250^2+1im,logger=1,tol=1e-10,errmeasure=RefErrmeasure);
Iteration:  1 errmeasure:1.274091618457501296e-01
...
Iteration: 49 errmeasure:1.269396691930517923e-10
Iteration: 50 errmeasure:7.608430406801784718e-11

As a NEP-solver developer

NEP-solvers should use the Errmeasure as follows. The NEP-solver should take as input a ErrmeasureType. This corresponds to either a function or a type, but if you follow the procedure below, you will not have to worry about that.

function mysolver(nep::NEP;errmeasure::ErrmeasureType=DefaultErrmeasure)

Before the main iteration, you need to initialize the error measure computation. The precomptued data is stored in a variable typically called ermdata:

   ermdata=init_errmeasure(errmeasure,nep);

In the main for loop you want to call the estimate_error function:

for k=1:maxit
    err=estimate_error(ermdata,λ,v)
    if (err < 1e-10)
       return (λ,v)
    end
    ....

end

Methods and types

NonlinearEigenproblems.NEPTypes.Errmeasure — Type.

abstract type Errmeasure; end

Concrete subtypes of Errmeasure represent specific ways of measuring the error of an eigenpair. NEP-solvers take a type as input and then instantiate a Errmeasure. As a NEP-solver user, you use the type as follows

julia> quasinewton(nep,errmeasure=ResidualErrmeasure)

User-specified ways of measuring error can be given by creating a new subtype of Errmeasure and using it as a errmeasure keyword. You need to specify the way to measure the error in the method estimate_error and (optionally) init_errmeasure.

Example

This shows how to compute a reference solution and then use this as a reference solution. The error in the second run will be effectively the eigenvalue error.

julia> nep=nep_gallery("qdep0");
julia> (λref,vref)=quasinewton(nep,λ=-1);
julia> struct EigvalError <: Errmeasure; nep::NEP; end
julia> function NonlinearEigenproblems.estimate_error(E::EigvalError,λ,v)
return abs(λref-λ);
end
julia> (λ,v)=quasinewton(nep,errmeasure=EigvalError,λ=-1.0 ,logger=1,tol=5e-13)
Precomputing linsolver
Iteration:  1 errmeasure:2.466988587467300320e-03, λ=-1.0 + 0.0im
Iteration:  2 errmeasure:4.625160667012763183e-01, λ=-0.539950921886191 + 0.0im
Iteration:  3 errmeasure:2.799848726755167494e-01, λ=-1.282451861262984 + 0.0im
Iteration:  4 errmeasure:3.422925625951256379e-02, λ=-1.0366962448469799 + 0.0im
Iteration:  5 errmeasure:5.530128437585268841e-04, λ=-1.0019139757437088 + 0.0im
Iteration:  6 errmeasure:1.159388768512403800e-04, λ=-1.002351049710616 + 0.0im
Iteration:  7 errmeasure:2.658434455016234210e-06, λ=-1.0024643301530123 + 0.0im
Iteration:  8 errmeasure:1.726871190488310503e-07, λ=-1.0024668159003483 + 0.0im
Iteration:  9 errmeasure:4.819693533164581822e-09, λ=-1.0024669934071608 + 0.0im
Iteration: 10 errmeasure:5.234268574128009277e-10, λ=-1.0024669880640404 + 0.0im
Iteration: 11 errmeasure:3.762568034915148019e-11, λ=-1.002466988625093 + 0.0im
Iteration: 12 errmeasure:3.205657961302676995e-12, λ=-1.0024669885842616 + 0.0im
Iteration: 13 errmeasure:3.352873534367972752e-14, λ=-1.0024669885874338 + 0.0im

Note that this can also be achieved by providing a function handle:

julia> myerrmeasure= (λ,v) -> abs(λref-λ);
julia> (λ,v)=quasinewton(nep,errmeasure=myerrmeasure,λ=-1.0 ,logger=1,tol=5e-13)
...
Iteration: 12 errmeasure:3.205657961302676995e-12, λ=-1.0024669885842616 + 0.0im
Iteration: 13 errmeasure:3.352873534367972752e-14, λ=-1.0024669885874338 + 0.0im

See also: DefaultErrmeasure, ResidualErrmeasure, BackwardErrmeasure, estimate_error, init_errmeasure.

NonlinearEigenproblems.NEPTypes.estimate_error — Function.

function estimate_error(E::Errmeasure,λ,v)

Returns the error estimate for the eigenpair (λ,v). The way to measure the error is specified in Errmeasure.

See also: Errmeasure

NonlinearEigenproblems.NEPTypes.init_errmeasure — Function.

function init_errmeasure(E::Errmeasure,nep)

This function is called in a precomputation phase for an error measure. For user defined Errmeasure, you do not need to overload this if your Errmeasure contains only one field which is a NEP.

See also: Errmeasure

NonlinearEigenproblems.NEPTypes.DefaultErrmeasure — Type.

struct DefaultErrmeasure <: Errmeasure

When you specify this Errmeasure, NEP-PACK tries to determine a suitable Errmeasure based on the type of the NEP. Note that this behavior may change in future versions.

See also: Errmeasure

NonlinearEigenproblems.NEPTypes.ResidualErrmeasure — Type.

struct ResidualErrmeasure <: Errmeasure

This Errmeasure species that the residual norm should be used to measure the error.

See also: Errmeasure

NonlinearEigenproblems.NEPTypes.BackwardErrmeasure — Type.

struct BackwardErrmeasure <: Errmeasure

This Errmeasure provides a way to compute the backward error. The backward error estimate are only given for NEPs which are subtypes of AbstractSPMF.

Example

julia> nep=nep_gallery("qdep0");
julia> (λ,v)=quasinewton(nep,λ=-1,errmeasure=BackwardErrmeasure,tol=1e-10);

See also: Errmeasure