Inner solvers

Inner solvers

Many NEP-solvers are based on a computation of a projected problem, i.e., if $V,W\in\mathbb{R}^{n\times p}$ we need to solve the (smaller) NEP

\[W^HM(λ)Vz=0\]

This is sometimes called a nonlinear Rayleigh-Ritz procedure, or a direct projection.

NEP-PACK provides a framework to handle projected problems. You can in principle use any of the NEP-solvers to solve a projected problem. As a user, this is specified in the inner_solver_method keyword argument.

If you wish to use the infinite Arnoldi method to handle the project solves in the nonlinear Arnoldi method, you can do the following:

julia> nep=nep_gallery("dep0_tridiag");
julia> λ,v=nlar(nep,neigs=1,inner_solver_method=IARInnerSolver(),logger=1);
Using inner solver IARInnerSolver(1.0e-13, 80, :ones, false, NonlinearEigenproblems.NEPSolver.iar)
iter 1 err:0.05095382004494062 λ=0.7579134426195271 - 0.03707164055891316im
iter 2 err:0.00031997693290503965 λ=-0.00010049358638757657 + 0.0001763732030940319im
iter 3 err:6.563177508431498e-6 λ=-0.0005335154073888051 - 4.498082881742902e-6im
iter 4 err:8.037612383023366e-9 λ=-0.0005259618179586685 + 2.806438064753968e-9im
iter 5 err:3.386041718599221e-11 λ=-0.0005259683064505526 + 1.3119844096548209e-11im
iter 6 err:3.4499779767886924e-13 λ=-0.0005259682927702825 - 1.0404412824030558e-13im
iter 7 err:5.365662696809372e-15 λ=-0.0005259682929434785 - 2.0247938561528697e-17im
****** 1 converged to eigenvalue: -0.0005259682929434785 - 2.0247938561528697e-17im errmeasure:5.365662696809372e-15

The logging of the inner solver is controlled by the kwarg inner_logger. This produces very verbose output illustrating also the convergence of the inner solve:

julia> λ,v=nlar(nep,neigs=1,inner_solver_method=IARInnerSolver(),logger=1,inner_logger=1);
Using inner solver IARInnerSolver(1.0e-13, 80, :ones, false, NonlinearEigenproblems.NEPSolver.iar)
-
--
---
----
-----
------
-------
--------
---------
----------
-----------
------------
=------------
+-------------
iter 1 err:0.04860520921206162 λ=0.8437634284420165 + 0.005742468974957178im
-
--
---
+---
iter 2 err:0.0008771218464072076 λ=-0.00065275394814732 - 0.0008601482370586537im
-
--
---
...

Inner solvers

The inner solvers inherit from InnerSolver. The following inner solvers are available by default.

NonlinearEigenproblems.NEPSolver.NewtonInnerSolver — Type.

struct NewtonInnerSolver
function NewtonInnerSolver(;tol=1e-13,maxit=80,starting_vector=:Vk,
                           newton_function=augnewton)

Uses a Newton-like method to solve the inner problem, with tolerance, and maximum number of iterations given by tol and maxit. The starting vector can be :ones, :randn, or :Vk. The value :Vk specifies the use of the super NEP-solver keyword argument (Vk). This is typically the previous iterate in the outer method.

The kwarg newton_function, specifies a Function which is called. We support augnewton, newton, resinv quasinewton, newtonqr. In principle it can be any method which takes the same keyword arguments as these methods.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.IARInnerSolver — Type.

struct IARInnerSolver
IARInnerSolver(;tol=1e-13,maxit=80,starting_vector=:ones,normalize_DEPs=:auto)

Uses iar to solve the inner problem, with tolerance, and maximum number of iterations given by tol and maxit. The starting vector can be :ones or :randn. normalize_DEPs determines if the we should carry out precomputation of DEPs (can speed up performance). It can take the value true, false or :auto. :auto sets it to true if we use the iar_chebyshev solver.

The kwarg iar_function, specifies a Function which is called. Examples of functions are iar and iar_chebyshev. It can be any NEP-solver which takes the same keyword arguments as these methods.

See also: InnerSolver, inner_solve

Missing docstring.

Missing docstring for IARChebshevInnerSolver. Check Documenter's build log for details.

NonlinearEigenproblems.NEPSolver.ContourBeynInnerSolver — Type.

struct ContourBeynInnerSolver <: InnerSolver

Uses contour_beyn to solve the inner problem, with radius and number of quadrature nodes, given by radius and n. If the variable radius is set to :auto, the integration radius will be automatically by using the eigenvalues approximations.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.PolyeigInnerSolver — Type.

struct PolyeigInnerSolver <: InnerSolver

For polynomial eigenvalue problems. Uses polyeig to solve the inner problem.

See also: InnerSolver, inner_solve

NonlinearEigenproblems.NEPSolver.SGIterInnerSolver — Type.

struct SGIterInnerSolver <: InnerSolver

Uses sgiter to solve the inner problem.

See also: InnerSolver, inner_solve

Advanced usage

You can define your own inner solver by inheriting from InnerSolver and implementing the function inner_solve. Since the inner_solve obtains information from the solver via keyword arguments, you need to end your method signature with kwargs...).

NonlinearEigenproblems.NEPSolver.InnerSolver — Type.

abstract type InnerSolver

Structs inheriting from this type are used to solve inner problems in an inner-outer iteration.

The InnerSolver objects are passed to the NEP-algorithms, which uses it to dispatch the correct version of the function inner_solve. Utilizes existing implementations of NEP-solvers and inner_solve acts as a wrapper to these.

Example

There is a DefaultInnerSolver that dispatches an inner solver based on the provided NEP. However, this example shows how you can force nlar to use the IARInnerSolver for a PEP.

julia> nep=nep_gallery("pep0", 100);
julia> λ,v = nlar(nep, inner_solver_method=NEPSolver.IARInnerSolver(), neigs=1, num_restart_ritz_vecs=1, maxit=70, tol=1e-8);
julia> norm(compute_Mlincomb(nep,λ[1],vec(v)))
8.68118417430353e-9

See also: inner_solve, DefaultInnerSolver, NewtonInnerSolver, PolyeigInnerSolver, IARInnerSolver, IARChebInnerSolver, SGIterInnerSolver, ContourBeynInnerSolver

NonlinearEigenproblems.NEPSolver.inner_solve — Function.

inner_solve(is::InnerSolver,T_arit,nep;kwargs...)

Solves the projected linear problem with solver specied with is. This is to be used as an inner solver in an inner-outer iteration. T specifies which method to use. The most common choice is DefaultInnersolver. The function returns (λv,V) where λv is an array of eigenvalues and V a matrix with corresponding vectors. The struct T_arit defines the arithmetics used in the outer iteration and should prefereably also be used in the inner iteration.

Different inner_solve methods take different kwargs. These are standardized kwargs:

neigs: Number of wanted eigenvalues (but less or more may be returned)

σ: target specifying where eigenvalues

λv, V: Vector/matrix of guesses to be used as starting values

j: the jth eigenvalue in a min-max characterization

tol: Termination tolarance for inner solver

inner_logger: Determines how the inner solves are logged. See Logger for further references