Logger

Logger

This is description of the logger functionality in NEP-PACK.

** TODO **

As a user

NonlinearEigenproblems.NEPSolver.Logger — Type.

abstract type Logger ; end

The type represents a way to log information throughout the algorithms. Error history and other properties can be stored in the logger. The most common Loggers are PrintLogger and ErrorLogger.

As a method developer you want to use push_info! and push_iteration_info!

See also: PrintLogger and ErrorLogger.

NonlinearEigenproblems.NEPSolver.PrintLogger — Type.

struct PrintLogger <: Logger ;
function PrintLogger(displaylevel)

When you use this logger, you will obtain printouts in stdout, no other logging. The displaylevel parameter specifies how much should be printed. The higher the value, the more is printed. Zero means no printouts.

NonlinearEigenproblems.NEPSolver.ErrorLogger — Type.

struct ErrorLogger <: Logger
ErrorLogger(nof_eigvals=100,nof_iters=100,displaylevel=1)

Use this object if you want to save the error history in a method. The displaylevel is interpreted as in PrintLogger. The kwargs nof_eigvals and nof_iters is used to preallocate memory used for saving.

When you use this logger, the error of push_iteration_info!-calls will be stored in logger.errs.

As a solver developer

NonlinearEigenproblems.NEPSolver.push_info! — Function.

function push_info!(logger, [level,] v; continues::Bool=false)

Pushes a string v to the logger. If continues=true, the next push_info! (or push_iteration_info!) is connected to this, e.g. line-feed will be omitted.

NonlinearEigenproblems.NEPSolver.push_iteration_info! — Function.

function push_iteration_info!(logger, [level,] iter; kwargs)

Pushes information about a specific iteration iter. Standardized keywords are λ, err and v.