Transformations

Transformations

Due to the object oriented way of handle NEPs in NEP-PACK, a NEP-object can be transformed to another NEP-object in a number of ways. There is support for:

variable transformations,

expansions, and

deflation.

Change of variables

NonlinearEigenproblems.NEPTypes.shift_and_scale — Function.

shift_and_scale(orgnep::NEP;shift=0,scale=1)

Transforms the orgnep by defining a new NEP from the relation T(λ)=M(scale * λ+shift) where M is the orgnep. This function tries to preserve the NEP type, e.g., a shiftandscale operation on an SPMF-object, return an SPMF object. If it cannot preserve the type, it will return a nep of the struct ShiftScaledNEP.

Example

julia> nep0=nep_gallery("pep0")
julia> σ=3; α=10;
julia> nep1=shift_and_scale(nep0,shift=σ,scale=α)
julia> opnorm(compute_Mder(nep0,α*(4+4im)+σ)-compute_Mder(nep1,4+4im))
8.875435870738592e-12

NonlinearEigenproblems.NEPTypes.mobius_transform — Function.

mobius_transform(orgnep::NEP,[,a=1][,b=0][,c=0][,d=1])

Transforms a nep (orgnep) $M(λ)v$ to a new nep $T(λ)=M((aλ+b)/(cλ+d))$. This function tries to preserve the type such that T and M are of the same NEP-type (see shift_and_scale()). If it cannot be preserved it will return a MobiusTransformedNEP. It is in general advised to try to preserve the type, and the use of MobiusTransformedNEP can considerably slow down NEP-access.

Example

julia> nep0=nep_gallery("pep0")
julia> a=1; b=3; c=4; d=5;
julia> nep1=mobius_transform(nep0,a=a,b=b,c=c,d=d);
julia> s=3;
julia> opnorm(compute_Mder(nep0,(a*s+b)/(c*s+d))-compute_Mder(nep1,s))
0.0

Expansions

Missing docstring.

Missing docstring for taylor_expansion_pep. Check Documenter's build log for details.

NonlinearEigenproblems.NEPTypes.interpolate — Function.

interpolate([T=ComplexF64,] nep::NEP, intpoints::Array)

Interpolates a NEP in the points intpoints and returns a PEP, i.e., a polynomial eigenvalue problem in a monomial basis. See ChebPEP for Chebyshev interpolation. The optional argument T is the type in which the matrices of the PEP should be defined.

See also ChebPEP.

Deflation

A NEP can be transformed to another NEP by extending the problem in a way that it essentially removes eigenvalues. This type of deflation is described on the manual page for deflation.