NEP Transformations

Transforming NEPs

There are various ways to transform NEPs into other NEPs. The simplest example is the function shift_and_scale().

NonlinearEigenproblems.NEPTypes.shift_and_scale — Function.

shift_and_scale(orgnep::NEP;shift=0,scale=1)

Transforms the orgnep by defining a new NEP from the relation T(λ)=M(scale * λ+shift) where M is the orgnep. This function tries to preserve the NEP type, e.g., a shiftandscale operation on an SPMF-object, return an SPMF object. If it cannot preserve the type, it will return a nep of the struct ShiftScaledNEP.

Example

julia> nep0=nep_gallery("pep0")
julia> σ=3; α=10;
julia> nep1=shift_and_scale(nep0,shift=σ,scale=α)
julia> opnorm(compute_Mder(nep0,α*(4+4im)+σ)-compute_Mder(nep1,4+4im))
8.875435870738592e-12

source

Similarly mobius_transform() is more general than shift_and_scale which transform the problem using a Möbius transformation. The function taylor_exp create new PEP by doing truncating a Taylor expansion.

Projection

Several methods for NEPs are based on forming a smaller NEP, which we will refer to as a projection:

\[N(λ)=W^HM(λ)V,\]

where $V,W\in\mathbb{C}^{n\times p}$ and the corresponding projected problem

\[N(λ)u=0.\]

Types

NEPs for which this projection can be computed inherit from ProjectableNEP.

NonlinearEigenproblems.NEPTypes.ProjectableNEP — Type.

abstract ProjectableNEP <: NEP

A ProjectableNEP is a NEP which can be projected, i.e., one can construct the problem $W'*M(λ)Vw=0$ with the Proj_NEP. A NEP which is of this must have the function create_proj_NEP(orgnep::ProjectableNEP) implemented. This function must return a Proj_NEP.

See also: set_projectmatrices!.

Example:

julia> nep=nep_gallery("dep0");
julia> typeof(nep)
DEP{Float64,Array{Float64,2}}
julia> isa(nep,ProjectableNEP)
true
julia> projnep=create_proj_NEP(nep);
julia> e1 = Matrix(1.0*I,size(nep,1),1);
julia> set_projectmatrices!(projnep,e1,e1);
julia> compute_Mder(nep,3.0)[1,1]
-2.315345215259418
julia> compute_Mder(projnep,3.0)
1×1 Array{Float64,2}:
 -2.315345215259418

source

The result of the projection is represented in a Proj_NEP.

NonlinearEigenproblems.NEPTypes.Proj_NEP — Type.

abstract type Proj_NEP <: NEP

Proj_NEP represents a projected NEP. The projection is defined as the NEP

\[N(λ)=W^HM(λ)V\]

where $M(λ)$ is a base NEP and W and V rectangular matrices representing a basis of the projection spaces. Instances are created with create_proj_NEP. See create_proj_NEP for examples.

Any Proj_NEP needs to implement two functions to manipulate the projection:

set_projectmatrices!: Set matrices W and V

expand_projectmatrices!: Effectively expand the matrices W and V with one column.

source

One explicit instance is the Proj_SPMF_NEP.

NonlinearEigenproblems.NEPTypes.Proj_SPMF_NEP — Type.

struct Proj_SPMF_NEP <: Proj_NEP

This type represents the (generic) way to project NEPs which are AbstractSPMF. See examples in create_proj_NEP.

source

Associated functions

You can create a projected NEP with create_proj_NEP:

NonlinearEigenproblems.NEPTypes.create_proj_NEP — Function.

pnep=create_proj_NEP(orgnep::ProjectableNEP[,maxsize [,T]])

Create a NEP representing a projected problem $N(λ)=W^HM(λ)V$, where the base NEP is represented by orgnep. The optional parameter maxsize::Int determines how large the projected problem can be and T is the Number type used for the projection matrices (default ComplexF64). These are needed for memory preallocation reasons. Use set_projectmatrices! and expand_projectmatrices! to specify projection matrices $V$ and $W$.

Example:

The following example illustrates that a projection of a NEP is also a NEP and we can for instance call compute_Mder on it:

julia> nep=nep_gallery("pep0")
julia> V=Matrix(1.0*I,size(nep,1),2);
julia> W=Matrix(1.0*I,size(nep,1),2);
julia> pnep=create_proj_NEP(nep);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,3.0)
2×2 Array{Complex{Float64},2}:
 -2.03662+0.0im   13.9777+0.0im
 -1.35069+0.0im  -13.0975+0.0im
julia> W'*compute_Mder(nep,3.0)*V  # Gives the same result
2×2 Array{Float64,2}:
 -2.03662   13.9777
 -1.35069  -13.0975

If you know that you will only use real projection matrices, you can specify this in at the creation:

julia> pnep=create_proj_NEP(nep,2,Float64);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,3.0)
2×2 Array{Float64,2}:
 -2.03662   13.9777
 -1.35069  -13.0975

source

NonlinearEigenproblems.NEPTypes.set_projectmatrices! — Function.

set_projectmatrices!(pnep::Proj_NEP,W,V)

Set the projection matrices for the NEP to W and V, i.e., corresponding the NEP: $N(λ)=W^HM(λ)V$. See also create_proj_NEP.

Example

This illustrates if W and V are vectors of ones, the projected problem becomes the sum of the rows and columns of the original NEP.

julia> nep=nep_gallery("pep0")
julia> pnep=create_proj_NEP(nep);
julia> V=ones(200,1);  W=ones(200,1);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,0)
1×1 Array{Complex{Float64},2}:
 48.948104019482756 + 0.0im
julia> sum(compute_Mder(nep,0),dims=[1,2])
1×1 Array{Float64,2}:
 48.948104019482955

source

NonlinearEigenproblems.NEPTypes.expand_projectmatrices! — Function.

expand_projectmatrices!(nep::Proj_SPMF_NEP, Wnew, Vnew)

The projected NEP is updated by adding the last column of Wnew and Vnew to the basis. Note that Wnew and Vnew contain also the "old" basis vectors. See also create_proj_NEP

Example:

In the following example you see that the expanded projected problem has one row and column more, and the leading subblock is the same as the smaller projected NEP.

julia> nep=nep_gallery("pep0"); n=size(nep,1);
julia> V=Matrix(1.0*I,n,2); W=Matrix(1.0*I,n,2);
julia> pnep=create_proj_NEP(nep);
julia> set_projectmatrices!(pnep,W,V);
julia> compute_Mder(pnep,0)
2×2 Array{Complex{Float64},2}:
 0.679107+0.0im   -0.50376+0.0im
 0.828413+0.0im  0.0646768+0.0im
julia> Vnew=[V ones(n)]
julia> Wnew=[W ones(n)]
julia> expand_projectmatrices!(pnep,Wnew,Vnew);
julia> compute_Mder(pnep,0)
3×3 Array{Complex{Float64},2}:
 0.679107+0.0im   -0.50376+0.0im  -12.1418+0.0im
 0.828413+0.0im  0.0646768+0.0im   16.3126+0.0im
 -17.1619+0.0im   -10.1628+0.0im   48.9481+0.0im

source

Deflation

Due to structure of the representation of NEPs in NEP-PACK it is possible to do deflation, by transformation of the NEP-object. The deflation is based on theory provided in Effenbergers thesis and the main function consists of effenberger_deflation.

NonlinearEigenproblems.NEPTypes.effenberger_deflation — Function.

effenberger_deflation(nep::NEP,S0,V0)

This function creates a deflated NEP based on (S0,V0), which are assumed to an invariant pair of nep. Effectively, the function should return a NEP which has the same solutions as orgnep, except those corresponding to (S0,V0).

Example:

julia> nep=nep_gallery("dep0");
julia> (λ,v)=newton(nep);
julia> n=size(nep,1);
julia> S0=reshape([λ],1,1);
julia> V0=reshape(v,n,1);
julia> dnep=effenberger_deflation(nep,S0,V0)
julia> (λ2,v2)=augnewton(dnep);  # this converges to different eigval
julia> minimum(svdvals(compute_Mder(nep,λ2)))
9.323003321058995e-17

References

C. Effenberger, Robust solution methods for nonlinear eigenvalue problems, PhD thesis, 2013, EPF Lausanne

source