NEP Types

NEPTypes

The basic type

The basic class is the abstract class NEP which represents a NEP. All other defined NEPs should inherit from NEP, or from a more specialized version; see, e.g., ProjectableNEP or AbstractSPMF.

NonlinearEigenproblems.NEPCore.NEP — Type.

abstract NEP

A NEP object represents a nonlinear eigenvalue problem. All NEPs should implement

size(nep::NEP,d)

and at least one of the following

M = compute_Mder(nep::NEP,λ::Number,i::Integer=0)

V = compute_Mlincomb!(nep::NEP,λ::Number,V::AbstractVecOrMat,a::Vector)

MM = compute_MM(nep::NEP,S,V)

source

Below we list the most common types built-in to NEP-PACK, and further down how you can access the NEP. However, the structure is made for extendability, and hence it is possible for you to extend with your own class of NEPs.

SPMF

One of the most common problem types is the SPMF_NEP. SPMF is short for Sum of Products of Matrices and Functions and the NEP is described by

\[M(λ) = \sum_{i} A_i f_i(λ).\]

The constructor of the SPMF_NEP, takes the the matrices and the functions, but also a number of other (optional) parameters which may increase performance or preserve underlying types.

NonlinearEigenproblems.NEPTypes.SPMF_NEP — Method.

 SPMF_NEP(AA, fii, check_consistency, Schur_fact = false, align_sparsity_patterns = false, , Ftype)

Creates a SPMF_NEP consisting of matrices AA and functions fii. The SPMF_NEP is defined by a sum of products of matrices and functions

\[M(λ)=∑_i A_i f_i(λ).\]

All of the matrices $A_0,...$ are of size $n×n$ and $f_i$ are a functions. The functions $f_i$ must be defined for matrices in the standard matrix function sense.

Parameters

AA is a Vector of matrices. The matrices have to be of the same type. If you need a NEP with different types you can use SumNEP to construct a sum of two SPMF_NEP.

fii is a Vector of functions. Each function takes one parameter S. The functions must be available both as a scalar valid function and a matrix function. If S is a square matrix, fii[k](S) musst also be a square matrix. If S is a scalar fii[k](S) is a scalar.

check_consistency (default true) determines if we should initiate by running tests to verify that the fii satisfies the conditions that every function is valid both for matrices and scalars. This is done by using @code_typed and the functions need to be type-stable in that sense.

align_sparsity_patterns (default false) has effect only for sparse matrices (SparseMatrixCSC). If align_sparsity_patterns=true the SparseMatrixCSC matrices will be replaced by equivalent SparseMatrixCSC matrices where the colptr and rowval are identical. This increases the speed of some functions, e.g., compute_Mder. If align_sparsity_patterns=true the matrices in the NEP should be considered read only. If the sparsity patterns are completely or mostly distinct, it may be more efficient to set this flag to false.

Ftype (default ComplexF64) determines an underlying type of the functions. The output of any function should be "smaller" than the promoted type of the input and Ftype. More precisely, if F=fii[k], then the type logic is as follows eltype(F(λ))=promote_type(eltype(λ),Ftype).

Schur_fact (default false) determines if the compute_MM function should tridiagonalize the matrix before carrying out the computation. This can be faster for large matrices.

Example

julia> A0=[1 3; 4 5]; A1=[3 4; 5 6];
julia> id_op=S -> one(S) # Note: We use one(S) to be valid both for matrices and scalars
julia> exp_op=S -> exp(S)
julia> nep=SPMF_NEP([A0,A1],[id_op,exp_op]);
julia> compute_Mder(nep,1)-(A0+A1*exp(1))
2×2 Array{Float64,2}:
 0.0  0.0
 0.0  0.0

source

Abstract SPMFs

Many problems can be described in the class of SPMF. There might be more specialized and efficient implementations such as, e.g. PEP, DEP or REP. However, on an abstract level it may still be important to recognize the similarities. Hence there is an abstract class AbstractSPMF, which in itself inherits from ProjectableNEP.

NonlinearEigenproblems.NEPTypes.AbstractSPMF — Type.

abstract  AbstractSPMF <: ProjectableNEP

An AbstractSPMF is an abstract class representing NEPs which can be represented as a Sum of products of matrices and functions $M(λ)=Σ_i A_i f_i(λ)$, where i = 0,1,2,..., all of the matrices are of size n times n and f_i are functions.

Any AbstractSPMF has to have implementations of get_Av() and get_fv() which return the functions and matrices.

source

NonlinearEigenproblems.NEPTypes.get_Av — Function.

get_Av(nep::AbstractSPMF)

Returns an array of matrices $A_i$ in the AbstractSPMF: $M(λ)=Σ_i A_i f_i(λ)$

source

NonlinearEigenproblems.NEPTypes.get_fv — Function.

get_Av(nep::AbstractSPMF)

Returns an Array of functions (that can be evaluated both as scalar and matrix functions) $f_i$ in the AbstractSPMF: $M(λ)=Σ_i A_i f_i(λ)$

source

PEP

The Polynomial Eigenvalue Problem is described by

\[M(λ) = \sum_{i} A_i λ^i.\]

NonlinearEigenproblems.NEPTypes.PEP — Type.

struct PEP <: AbstractSPMF

A polynomial eigenvalue problem (PEP) is defined by the sum the sum $Σ_i A_i λ^i$, where i = 0,1,2,..., and all of the matrices are of size n times n.

source

In order to construct a PEP, we only need to provide the matrices.

NonlinearEigenproblems.NEPTypes.PEP — Method.

PEP(AA::Array)

Creates a polynomial eigenvalue problem with monomial matrices specified in AA, which is an array of matrices.

julia> A0=[1 3; 4 5]; A1=A0.+one(2); A2=ones(2,2);
julia> pep=PEP([A0,A1,A2])
julia> compute_Mder(pep,3)-(A0+A1*3+A2*9)
2×2 Array{Float64,2}:
 0.0  0.0
 0.0  0.0

source

DEP

The Delay Eigenvalue Problem is described by

\[M(λ) = -λI + \sum_{i} A_i e^{-τ_i λ}.\]

NonlinearEigenproblems.NEPTypes.DEP — Type.

type DEP <: AbstractSPMF

A DEP (Delay Eigenvalue problem) is defined by the sum $-λI + Σ_i A_i exp(-tau_i λ)$ where all of the matrices are of size n times n.
Constructor: DEP(AA,tauv) where AA is an array of the matrices $A_i$, and tauv is a vector of the values $tau_i$.

Example:

julia> A0=randn(3,3); A1=randn(3,3);
julia> tauv=[0,0.2] # Vector with delays
julia> dep=DEP([A0,A1],tauv)
julia> λ=3.0;
julia> M1=compute_Mder(dep,λ)
julia> M2=-λ*I+A0+A1*exp(-tauv[2]*λ)
julia> norm(M1-M2)
0.0

source

REP

The Rational Eigenvalue Problem is described by:

NonlinearEigenproblems.NEPTypes.REP — Type.

struct REP <: AbstractSPMF

A REP represents a rational eigenvalue problem. The REP is defined by the sum $Σ_i A_i s_i(λ)/q_i(λ)$, where i = 0,1,2,..., all of the matrices are of size n times n and si and qi are polynomials.

source

The constructor is called as:

NonlinearEigenproblems.NEPTypes.REP — Method.

REP(A,poles)

Creates a rational eigenvalue problem. The constructor takes the matrices A_i and a sequence of poles as input (not complete).

Example

julia> A0=[1 2; 3 4]; A1=[3 4; 5 6];
julia> nep=REP([A0,A1],[1,3]);
julia> compute_Mder(nep,3)
2×2 Array{Float64,2}:
 NaN  NaN
 NaN  NaN

source

SumNEP

It is also possible to consider NEPs that are summs of other NEPs. For such situations there are SumNEPs. Specifically GenericSumNEP and SPMFSumNEP. Both are constructed using the function SumNEP.

NonlinearEigenproblems.NEPTypes.SumNEP — Function.

SumNEP{nep1::NEP,nep2::NEP}
SumNEP{nep1::AbstractSPMF,nep2::AbstractSPMF}

SumNEP is a function creating an object that corresponds to a sum of two NEPs, i.e., if nep is created by SumNEP it is defined by

\[M(λ)=M_1(λ)+M_2(λ)\]

where M1 and M2 are defined by nep1 and nep2.

Example:

julia> nep1=DEP([ones(3,3),randn(3,3)])
julia> nep2=PEP([ones(3,3),randn(3,3),randn(3,3)])
julia> sumnep=SumNEP(nep1,nep2);
julia> s=3.0;
julia> M=compute_Mder(sumnep,s);
3×3 Array{Float64,2}:
  8.54014     6.71897   7.12007
 -0.943908  -13.0795   -0.621659
  6.03155    -7.26726  -6.42828
julia> M1=compute_Mder(nep1,s);
julia> M2=compute_Mder(nep2,s);
julia> M1+M2  # Same as M
3×3 Array{Float64,2}:
  8.54014     6.71897   7.12007
 -0.943908  -13.0795   -0.621659
  6.03155    -7.26726  -6.42828

See also: SPMFSumNEP, GenericSumNEP

source

NonlinearEigenproblems.NEPTypes.GenericSumNEP — Type.

struct GenericSumNEP{NEP1<:NEP,NEP2<:NEP}  <: NEP

See also: SumNEP, SPMFSumNEP

source

NonlinearEigenproblems.NEPTypes.SPMFSumNEP — Type.

struct SPMFSumNEP{NEP1<:AbstractSPMF,NEP2<:AbstractSPMF}  <: AbstractSPMF{AbstractMatrix}

See also: SumNEP, GenericSumNEP

source

Accessing the NEP

The nonlinear eigenvalue problem is defined by the data stored in the NEP-class, and the NEP-solvers access the data mainly through three main functions, compute_Mder compute_Mlincomb and compute_MM.

NonlinearEigenproblems.NEPCore.compute_Mder — Function.

compute_Mder(nep::NEP,λ::Number [,i::Integer=0])

Computes the ith derivative of nep evaluated in λ.

Example

This example shows that compute_Mder(nep,λ,1) gives the first derivative.

julia> nep=nep_gallery("dep0");
julia> ϵ=1e-5;
julia> Aminus=compute_Mder(nep,λ-ϵ);
julia> Aminus=compute_Mder(nep,λ-ϵ);
julia> Aplus=compute_Mder(nep,λ+ϵ);
julia> opnorm((Aplus-Aminus)/(2ϵ)-compute_Mder(nep,λ,1))
1.990970375089371e-11

source

NonlinearEigenproblems.NEPCore.compute_Mlincomb! — Function.

compute_Mlincomb(nep::NEP,λ::Number,V, a::Vector=ones(size(V,2)), startder=0)
compute_Mlincomb!(nep::NEP,λ::Number,V, a::Vector=ones(size(V,2)), startder=0)

Computes the linear combination of derivatives
$Σ_i a_i M^{(i)}(λ) v_i$ starting from derivative startder. The function compute_Mlincomb! does the same but may modify the V matrix/array.

Example

This example shows that compute_Mder gives a result consistent with compute_Mlincomb. Note that compute_Mlincomb is in general faster since no matrix needs to be constructed.

julia> nep=nep_gallery("dep0");
julia> v=ones(size(nep,1)); λ=-1+1im;
julia> norm(compute_Mder(nep,λ,1)*v-compute_Mlincomb(nep,λ,hcat(v,v),[0,1]))
1.0778315928076987e-15

source

NonlinearEigenproblems.NEPCore.compute_MM — Function.

compute_MM(nep::NEP,S,V)

Computes the sum $Σ_i M_i V f_i(S)$ for a NEP, where $S$ and $V$ are matrices, and the NEP satisfies $M(λ)=Σ_i M_i f_i(λ)$.

Example

This example shows that for diagonal S, the result of compute_MM can also be computed with compute_Mlincomb

julia> nep=nep_gallery("dep0");
julia> D=diagm(0 => [1,2])
2×2 Array{Int64,2}:
 1  0
 0  2
julia> V=ones(size(n,1),2);
julia> W=compute_MM(nep,D,V);
julia> norm(W[:,1]-compute_Mlincomb(nep,D[1,1],V[:,1]))
1.1102230246251565e-16
julia> norm(W[:,2]-compute_Mlincomb(nep,D[2,2],V[:,2]))
0.0

Reference

Properties of the quantity $Σ_i M_i V f_i(S)$ for non-polynomial nonlinear eigenvalue problems were extensively used in:

D. Kressner A block Newton method for nonlinear eigenvalue problems, Numer. Math., 114 (2) (2009), pp. 355-372

C. Effenberger, Robust solution methods for nonlinear eigenvalue problems, PhD thesis, 2013, EPF Lausanne

source